water tap, as exported from unreal
The greatest possible water tap you could ever ask for. :P
for future ref: as tearing apart the merged mesh from unreal has proven almost impossible (losing details at each step) I am taking a shortcut, and adding the same mesh that is used for the sink, for the tap also. Both now produce the same visual, but the rest should be fine. also included the old visual in the folder, but it's not linked.